+++ netsuite/restlet.py
@@ -1,8 +1,8 @@
 import logging
+from functools import cached_property
 
 from . import rest_api_base
 from .config import Config
-from .util import cached_property
 
 logger = logging.getLogger(__name__)